requirements.txt
pandas==1.5.3
numpy==1.25.0
scikit-learn==1.2.2
opencv-python==4.7.0.72
pyfeats==1.0.1
pyradiomics==3.0.1
scipy==1.11.0
simpleitk==2.2.1
configargparse==1.4
pydicom==2.4.1
pyyaml==6.0
openpyxl==3.0.10
seaborn==0.12.2
boruta==0.3